move to models about models code
