+++ flaskr/models/post.py
@@ -1,0 +1,17 @@
+from datetime import datetime
+
+db = SQLAlchemy(app)
+
+class Post(db.Model):
+
+    __tablename__ = 'posts'
+
+    id = db.Column(db.Integer, primary_key=True)
+    user_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
+    created_at = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
+    updated_at = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
+    title = db.Column(db.Text, nullable=False)
+    body = db.Column(db.Text, nullable=False)
+
+    users = db.relationship('User', backref=db.backref('posts', lazy=True))
+
